Fixed last test, however the new functionality needs to be tested in more than the way that I fixed the old tests.
It's also probably worthwhile to find a better way to accomplish the information
propogation necessary to unit test the complex spread of things in this
simulator. I created a MockEventManager that will basically execute arbitrary
events in the proper order until there are none, however this won't really work
for a CallMarket so maybe not the best? I think we can do better.