Removed simulation name and number.
The log file merger will prepend each line with the simulation number, which is
determined by the order they are passed into the log file merger.

The simulation name is found using the distinct part of the absolute file
path. Thus, the output should be the same as before, but this will also be
correct if the simulations are in arbitrary folders.