Fundamental now only logs the fundamental at times it's queried, and "interpolates" for the times between. Observations based off of time series objects now appropriately handle for this sparsity by making the same interpolation assumptions as TimeSeries. Note that this interpolation is essentially hard coded as it's basically the only logical way to interpolate that allows us to easily compute statistics in linear time with the number of samples, not the length of time.
The interpolation assumes that the value in unchanged after the last
observation, which is logical in the case of spread and median, and is somewhat
biased in terms of the fundamental. However, when we're concerned about
variance, no interpolation will accurately account, without actually
estimating the variance which would be very involved for likely little
gain.

The sampling for periodic measures is still accurate under the assumptions of
the interpolation, which will probably result in more accurate proxies than
using the point-by-point or 1-period measures, especially for values like
variance.